Fix an issue where server side configs are ignored.
Fix an issue where server side configs are ignored if a .criterion.yml
config file does not exist. It will now take into account the server
side config too when detecting test type.

This allows us to set server side configs such as "setup" or "script"
for projects that do not contain .criterion.yml files. If neither are
set it will revert back to trying to work out if its a PHPUnit project
with composer etc.